visualization/requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
PyYAML==6.0
scipy==1.8.0
tabulate==0.8.9
